Move the App component to under the components dir
