+++ examples/shopping-cart/src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import ProductsContainer from './ProductsContainer'
-import CartContainer from './CartContainer'
+import ProductsContainer from '../containers/ProductsContainer'
+import CartContainer from '../containers/CartContainer'
 
 const App = () => (
   <div>
